move manifold tags to <CGAL/tags.h>
