+++ STL_Extension/doc/STL_Extension/CGAL/tags.h
@@ -102,3 +102,74 @@ struct Null_tag {
 }; /* end Null_tag */
 
 } /* end namespace CGAL */
+
+namespace CGAL {
+
+/*!
+\ingroup PkgSTLExtensionUtilities
+
+The class `Manifold_tag` is a tag class used to monitor the
+surface meshing algorithm. When instantiated with the tag
+`Manifold_tag` the function template
+`make_surface_mesh()`
+ensures that the output mesh is a manifold surface
+without boundary.
+
+\sa `make_surface_mesh()`
+\sa `Manifold_with_boundary_tag`
+\sa `Non_manifold_tag`
+
+*/
+
+struct Manifold_tag {
+
+}; /* end Manifold_tag */
+} /* end namespace CGAL */
+
+namespace CGAL {
+
+/*!
+\ingroup PkgSTLExtensionUtilities
+
+The class `Manifold_with_boundary_tag` is a tag class used to monitor the
+surface meshing algorithm. When instantiated with the tag
+`Manifold_with_boundary_tag`, the function template
+`make_surface_mesh()`
+ensures that the output mesh is a manifold surface
+but it may have boundaries.
+
+\sa `make_surface_mesh()`
+\sa `Manifold_tag`
+\sa `Non_manifold_tag`
+
+*/
+
+struct Manifold_with_boundary_tag {
+
+}; /* end Manifold_with_boundary_tag */
+} /* end namespace CGAL */
+
+namespace CGAL {
+
+/*!
+\ingroup PkgSTLExtensionUtilities
+
+The class `Non_manifold_tag` is a tag class used to monitor the
+surface meshing algorithm. When instantiated with the tag
+`Non_manifold_tag` the function template
+`make_surface_mesh()`
+does not ensure that the output mesh is a manifold surface.
+The manifold property of output mesh
+may nevertheless result from the choice of
+appropriate meshing criteria.
+
+\sa `make_surface_mesh()`
+\sa `Manifold_tag`
+\sa `Manifold_with_boundary_tag`
+
+*/
+
+struct Non_manifold_tag {
+
+}; /* end Non_manifold_tag */
+} /* end namespace CGAL */
