+++ STL_Extension/doc/STL_Extension/dependencies
@@ -4,4 +4,5 @@ Number_types
 Kernel_23
 Miscellany
 Surface_mesh
+Surface_mesher
 BGL
