+++ Surface_mesher/doc/Surface_mesher/dependencies
@@ -8,3 +8,4 @@ Triangulation_2
 Triangulation_3
 Mesh_3
 Polyhedron
+STL_Extension
