+++ Surface_mesher/include/CGAL/make_surface_mesh.h
@@ -19,6 +19,7 @@
   "The 3D Mesh Generation package (see https://doc.cgal.org/latest/Mesh_3/) should be used instead."
 #include <CGAL/Installation/internal/deprecation_warning.h>
 
+#include <CGAL/tags.h>
 #include <CGAL/Surface_mesher_generator.h>
 
 #include <CGAL/Surface_mesh_complex_2_in_triangulation_3.h>
